Add files via upload
The main Jupyter Notebook which contains the complete data exploration, pre-processing, and model(s) training.

2 datasets have been used in this source code: 1 is the images dataset and the other is the text dataset. The text dataset contains the text written on the images/memes.

We first work on the text dataset & models, where after having read the dataset, we apply the count vectorizer and then split the dataset into 5 categories: 
i) Very Positive     ii) Positive     iii) Neutral     iv) Negative     v) Very Negative

We combine "Very Positive" and "Positive" into 1 category, and "Negative" and "Very Negative" into 1 category (due to limited data, despite applying other techniques).
-> The model will predict one of these above categories. 

Having calculated the count of the 5 (initial) categories, we notice that the dataset is greatly imbalanced and with "Positive" being the most common category and "Very Negative" being the most rare category. To bring balance in this, over sampling and under sampling is done such that all the training categories are equal and all the testing categories are equal. This ensures that there's no skewness while training and that our models predict equally for all the 3 categories. 

Next we train 3 models for our text and then also evaluate them; K-Neighbours Classifier, Decision Tree, and a Linear Regression Model is trained and a confusion matrix and f1 score is calculated and printed for all 3 models. 

Moving on we now pre-process our images dataset. The images are first appended into a list and then sorted according to their filename, which in our case are numbers. Next we read the images, using the filename stored in the list, resize all the images to a standard size and then calculate the hog features of every image. These are appended into a dataframe along with the filename. Our images dataframe is also split into training and testing so that we can train the images model on this spllited data. 

For our images dataset, we again train and evaluate 3 models; K-Neighbours Classifier, Decision Tree, and a Linear Regression Model. Again, a confusion matrix and f1 score is calculated and printed for all 3 models. 

Lastly, we save all our models, for both text and images, using pickle so that we can re-use these in the future, when needed. We then apply a voting classifier on both text and images models, separately i.e 1 for the 3 text models, and 1 for the 3 images models. Their combined F1 scores are calculated and confusion matrixes are drawn/printed. 


*PLEASE NOTE THAT THIS PROJECT WAS COMPLETED DURING MY SOPHMORE YEAR (3rd Semester) IN UNIVERSITY I.E IN FALL 2022.